# backend/openapi_generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGRO",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/stock-entry/": {
      "post": {
        "tags": [
          "Stock Entry"
        ],
        "summary": "Create",
        "operationId": "create_v1_stock_entry__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Stock Entry"
        ],
        "summary": "List stock entries",
        "description": "Retrieve all stock entries with optional date filter.\n\nArgs:\n    date (Optional[date]): Filter by entry date.\n    skip (int): Number of records to skip.\n    limit (int): Maximum number of records to return.\n    db (Session): Database session dependency.\n\nReturns:\n    List[StockEntryRead]: List of stock entries.",
        "operationId": "read_all_v1_stock_entry__get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by date",
              "title": "Date"
            },
            "description": "Filter by date"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockEntryRead"
                  },
                  "title": "Response Read All V1 Stock Entry  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stock-entry/{stock_entry_id}": {
      "get": {
        "tags": [
          "Stock Entry"
        ],
        "summary": "Get stock entry by ID",
        "description": "Retrieve a single stock entry by ID.\n\nArgs:\n    stock_entry_id (int): Stock entry ID.\n    db (Session): Database session dependency.\n\nReturns:\n    StockEntryRead: The stock entry.\n\nRaises:\n    AppException: If entry not found (404).",
        "operationId": "read_one_v1_stock_entry__stock_entry_id__get",
        "parameters": [
          {
            "name": "stock_entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Stock Entry"
        ],
        "summary": "Update stock entry",
        "description": "Update an existing stock entry.\n\nArgs:\n    stock_entry_id (int): Stock entry ID.\n    entry_update (StockEntryUpdate): Update data.\n    db (Session): Database session dependency.\n\nReturns:\n    StockEntryRead: The updated entry.\n\nRaises:\n    AppException: If entry not found (404).",
        "operationId": "update_v1_stock_entry__stock_entry_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "stock_entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockEntryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Stock Entry"
        ],
        "summary": "Delete stock entry",
        "description": "Delete a stock entry by ID.\n\nArgs:\n    stock_entry_id (int): Stock entry ID.\n    db (Session): Database session dependency.\n\nRaises:\n    AppException: If entry not found (404).",
        "operationId": "delete_v1_stock_entry__stock_entry_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "stock_entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Entry Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/stock-entry/stock-entry/": {
      "post": {
        "tags": [
          "Stock Entry",
          "Stock Entry"
        ],
        "summary": "Create",
        "operationId": "create_v1_stock_entry_stock_entry__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Stock Entry",
          "Stock Entry"
        ],
        "summary": "List stock entries",
        "description": "Retrieve all stock entries with optional date filter.\n\nArgs:\n    date (Optional[date]): Filter by entry date.\n    skip (int): Number of records to skip.\n    limit (int): Maximum number of records to return.\n    db (Session): Database session dependency.\n\nReturns:\n    List[StockEntryRead]: List of stock entries.",
        "operationId": "read_all_v1_stock_entry_stock_entry__get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by date",
              "title": "Date"
            },
            "description": "Filter by date"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockEntryRead"
                  },
                  "title": "Response Read All V1 Stock Entry Stock Entry  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stock-entry/stock-entry/{stock_entry_id}": {
      "get": {
        "tags": [
          "Stock Entry",
          "Stock Entry"
        ],
        "summary": "Get stock entry by ID",
        "description": "Retrieve a single stock entry by ID.\n\nArgs:\n    stock_entry_id (int): Stock entry ID.\n    db (Session): Database session dependency.\n\nReturns:\n    StockEntryRead: The stock entry.\n\nRaises:\n    AppException: If entry not found (404).",
        "operationId": "read_one_v1_stock_entry_stock_entry__stock_entry_id__get",
        "parameters": [
          {
            "name": "stock_entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Stock Entry",
          "Stock Entry"
        ],
        "summary": "Update stock entry",
        "description": "Update an existing stock entry.\n\nArgs:\n    stock_entry_id (int): Stock entry ID.\n    entry_update (StockEntryUpdate): Update data.\n    db (Session): Database session dependency.\n\nReturns:\n    StockEntryRead: The updated entry.\n\nRaises:\n    AppException: If entry not found (404).",
        "operationId": "update_v1_stock_entry_stock_entry__stock_entry_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "stock_entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockEntryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Stock Entry",
          "Stock Entry"
        ],
        "summary": "Delete stock entry",
        "description": "Delete a stock entry by ID.\n\nArgs:\n    stock_entry_id (int): Stock entry ID.\n    db (Session): Database session dependency.\n\nRaises:\n    AppException: If entry not found (404).",
        "operationId": "delete_v1_stock_entry_stock_entry__stock_entry_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "stock_entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Entry Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "item_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Code"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "default_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Unit"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "ItemRead"
      },
...
      "StockEntryCreate": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "received_date": {
            "type": "string",
            "format": "date",
            "title": "Received Date"
          },
          "price_per_unit": {
            "type": "number",
            "title": "Price Per Unit"
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "received_date",
          "price_per_unit",
          "total_cost",
          "quantity",
          "unit"
        ],
        "title": "StockEntryCreate"
      },
      "StockEntryRead": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "received_date": {
            "type": "string",
            "format": "date",
            "title": "Received Date"
          },
          "price_per_unit": {
            "type": "number",
            "title": "Price Per Unit"
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "item": {
            "$ref": "#/components/schemas/ItemRead"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "received_date",
          "price_per_unit",
          "total_cost",
          "quantity",
          "unit",
          "id",
          "batch_id",
          "item",
          "created_at",
          "updated_at",
          "created_by",
          "updated_by"
        ],
        "title": "StockEntryRead"
      },
      "StockEntryUpdate": {
        "properties": {
          "received_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received Date"
          },
          "price_per_unit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Unit"
          },
          "total_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cost"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "type": "object",
        "title": "StockEntryUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}